project0: Add logic to only update the highest offer on a shop item of a customer's offer is higher than the highest offer

+++ projects/project0/src/main/java/com/revature/services/CustomerService.java
@@ -16,8 +16,8 @@ public class CustomerService {
 
 	private ShopItemDAO sid = new ShopItemPostgres();
 
-	public boolean makeOffer(double offer, ShopItem si) {
-		//
+	public boolean makeOffer(float offer, ShopItem si) {
+		
 		return sid.makeShopItemOffer(offer, si);
 	}
 }
